Start parameterizing modules
This means that modules are starting to be instantiated with their own
custom logger and persistence database. Functionally nothing has
changed for now, but this lays the groundwork for a lot of stuff in
the future.

At the same time, we change lib/app.js to return new Routers and
Express applications, instead of exporting single global versions. I
would have liked to do this separately but it was just so much easier
since the tests had to be refactored at the same time, due to the
parameterization. So whatever.